pyproject.toml
[project]
name = "discord-link-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "tortoise-orm>=0.21.0",
    "discord-pretty-help>=1.3.4",
    "python-dotenv>=1.2.1",
    "aws-cdk-lib",
    "constructs>=10.0.0,<11.0.0",
    "cryptography",
]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
    "pyright>=1.1.390",
    "ty>=0.0.1a26",
    "pylint>=4.0.2",
    "awsebcli>=3.25.3",
]

[tool]
